+++ test/src/unit-capi-config.cc
@@ -201,7 +201,6 @@ void check_save_to_file() {
   ss << "sm.consolidation.steps 4294967295\n";
   ss << "sm.dedup_coords false\n";
   ss << "sm.enable_signal_handlers true\n";
-  ss << "sm.fragment_metadata_cache_size 10000000\n";
   ss << "sm.memory_budget 5368709120\n";
   ss << "sm.memory_budget_var 10737418240\n";
   ss << "sm.num_async_threads 1\n";
@@ -375,8 +374,6 @@ TEST_CASE("C API: Test config iter", "[capi], [config]") {
   all_param_values["sm.tile_cache_size"] = "100";
   all_param_values["sm.memory_budget"] = "5368709120";
   all_param_values["sm.memory_budget_var"] = "10737418240";
-  all_param_values["sm.array_schema_cache_size"] = "1000";
-  all_param_values["sm.fragment_metadata_cache_size"] = "10000000";
   all_param_values["sm.enable_signal_handlers"] = "true";
   all_param_values["sm.num_async_threads"] = "1";
   all_param_values["sm.num_reader_threads"] = "1";