+++ tiledb/sm/misc/constants.h
@@ -261,9 +261,6 @@ extern const uint32_t var_num;
 /** String describing no compression. */
 extern const std::string no_compression_str;
 
-/** The fragment metadata cache size. */
-extern const uint64_t fragment_metadata_cache_size;
-
 /** Whether or not the signal handlers are installed. */
 extern const bool enable_signal_handlers;